Moved problems to specific sub-package
