+++ compiler-plugin/src/main/kotlin/tech/mappie/validation/problems/UnknownParameterNameProblems.kt
@@ -1,4 +1,4 @@
-package tech.mappie.validation
+package tech.mappie.validation.problems
 
 import org.jetbrains.kotlin.ir.declarations.IrConstructor
 import org.jetbrains.kotlin.ir.util.dumpKotlinLike
@@ -6,13 +6,13 @@ import org.jetbrains.kotlin.name.Name
 import tech.mappie.MappieIrRegistrar.Companion.context
 import tech.mappie.resolving.ConstructorCallMapping
 import tech.mappie.util.location
+import tech.mappie.validation.Problem
 
 class UnknownParameterNameProblems(private val unknowns: List<Name>) {
 
     fun all(): List<Problem> = unknowns.map {
         Problem.error("Parameter ${it.asString()} does not occur as a parameter in constructor")
     }
-
 
     companion object {
         fun of(mapping: ConstructorCallMapping): UnknownParameterNameProblems =
